Create provision_mail_server.sh
This script provisions a mail server with 1 and optionally 2 domains, to get you started and nearly done with the setup of a mail server with a GUI for webmail and another one to administrate the server; there you can add more users, domains and anything you want. 
It is meant to get the arguments from the ENV that is declared in the Dockerfile.