Create Dockerfile
Creates a mail server with 1 and optionally 2 domains, to get you started and nearly done with the setup of a mail server with a GUI for webmail and another one to administrate the server; there you can add more users, domains and anything you want easily later. Make sure you have setup the DNS (A, MX, TXT) records before you deploy it.